src/main/java/br/com/xbrain/autenticacao/modules/permissao/service/AplicacaoService.java
package br.com.xbrain.autenticacao.modules.permissao.service;

import br.com.xbrain.autenticacao.modules.permissao.model.Aplicacao;
import br.com.xbrain.autenticacao.modules.permissao.repository.AplicacaoRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;

@Service
public class AplicacaoService {
    @Autowired
    private AplicacaoRepository repository;

    public List<Aplicacao> getAll() {
        return repository.findAll();
    }
}
